+++ what4/src/What4/Solver/CVC5.hs
@@ -1,13 +1,13 @@
 ------------------------------------------------------------------------
 -- |
--- Module      : What4.Solver.CVC4
--- Description : Solver adapter code for CVC4
--- Copyright   : (c) Galois, Inc 2015-2020
+-- Module      : What4.Solver.CVC5
+-- Description : Solver adapter code for cvc5
+-- Copyright   : (c) Galois, Inc 2022
 -- License     : BSD3
 -- Maintainer  : Rob Dockins <rdockins@galois.com>
 -- Stability   : provisional
 --
--- CVC4-specific tweaks to the basic SMTLib2 solver interface.
+-- CVC5-specific tweaks to the basic SMTLib2 solver interface.
 ------------------------------------------------------------------------
 {-# LANGUAGE FlexibleInstances #-}
 {-# LANGUAGE GADTs #-}
@@ -15,17 +15,17 @@
 {-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE TypeApplications #-}
 
-module What4.Solver.CVC4
-  ( CVC4(..)
-  , cvc4Features
-  , cvc4Adapter
-  , cvc4Path
-  , cvc4Timeout
-  , cvc4Options
-  , runCVC4InOverride
-  , withCVC4
-  , writeCVC4SMT2File
-  , writeMultiAsmpCVC4SMT2File
+module What4.Solver.CVC5
+  ( CVC5(..)
+  , cvc5Features
+  , cvc5Adapter
+  , cvc5Path
+  , cvc5Timeout
+  , cvc5Options
+  , runCVC5InOverride
+  , withCVC5
+  , writeCVC5SMT2File
+  , writeMultiAsmpCVC5SMT2File
   ) where
 
 import           Control.Monad (forM_, when)
@@ -56,73 +56,60 @@ intWithRangeOpt :: ConfigOption BaseIntegerType -> Integer -> Integer -> ConfigD
 intWithRangeOpt nm lo hi = mkOpt nm sty Nothing Nothing
   where sty = integerWithRangeOptSty (Inclusive lo) (Inclusive hi)
 
-data CVC4 = CVC4 deriving Show
+data CVC5 = CVC5 deriving Show
 
--- | Path to cvc4
-cvc4Path :: ConfigOption (BaseStringType Unicode)
-cvc4Path = configOption knownRepr "solver.cvc4.path"
+-- | Path to cvc5
+cvc5Path :: ConfigOption (BaseStringType Unicode)
+cvc5Path = configOption knownRepr "solver.cvc5.path"
 
-cvc4PathOLD :: ConfigOption (BaseStringType Unicode)
-cvc4PathOLD = configOption knownRepr "cvc4_path"
-
-cvc4RandomSeed :: ConfigOption BaseIntegerType
-cvc4RandomSeed = configOption knownRepr "solver.cvc4.random-seed"
-
-cvc4RandomSeedOLD :: ConfigOption BaseIntegerType
-cvc4RandomSeedOLD = configOption knownRepr "cvc4.random-seed"
+cvc5RandomSeed :: ConfigOption BaseIntegerType
+cvc5RandomSeed = configOption knownRepr "solver.cvc5.random-seed"
 
 -- | Per-check timeout, in milliseconds (zero is none)
-cvc4Timeout :: ConfigOption BaseIntegerType
-cvc4Timeout = configOption knownRepr "solver.cvc4.timeout"
+cvc5Timeout :: ConfigOption BaseIntegerType
+cvc5Timeout = configOption knownRepr "solver.cvc5.timeout"
 
-cvc4TimeoutOLD :: ConfigOption BaseIntegerType
-cvc4TimeoutOLD = configOption knownRepr "cvc4_timeout"
-
--- | Control strict parsing for CVC4 solver responses (defaults
+-- | Control strict parsing for cvc5 solver responses (defaults
 -- to solver.strict-parsing option setting).
-cvc4StrictParsing :: ConfigOption BaseBoolType
-cvc4StrictParsing = configOption knownRepr "solver.cvc4.strict_parsing"
+cvc5StrictParsing :: ConfigOption BaseBoolType
+cvc5StrictParsing = configOption knownRepr "solver.cvc5.strict_parsing"
 
-cvc4Options :: [ConfigDesc]
-cvc4Options =
+cvc5Options :: [ConfigDesc]
+cvc5Options =
   let pathOpt co = mkOpt co
                    executablePathOptSty
-                   (Just "Path to CVC4 executable")
-                   (Just (ConcreteString "cvc4"))
-      p1 = pathOpt cvc4Path
-      r1 = intWithRangeOpt cvc4RandomSeed (negate (2^(30::Int)-1)) (2^(30::Int)-1)
+                   (Just "Path to CVC5 executable")
+                   (Just (ConcreteString "cvc5"))
+      p1 = pathOpt cvc5Path
+      r1 = intWithRangeOpt cvc5RandomSeed (negate (2^(30::Int)-1)) (2^(30::Int)-1)
       tmOpt co = mkOpt co
                  integerOptSty
                  (Just "Per-check timeout in milliseconds (zero is none)")
                  (Just (ConcreteInteger 0))
-      t1 = tmOpt cvc4Timeout
+      t1 = tmOpt cvc5Timeout
   in [ p1, r1, t1
-     , copyOpt (const $ configOptionText cvc4StrictParsing) strictSMTParseOpt
-     , deprecatedOpt [p1] $ pathOpt cvc4PathOLD
-     , deprecatedOpt [r1] $ intWithRangeOpt cvc4RandomSeedOLD
-       (negate (2^(30::Int)-1)) (2^(30::Int)-1)
-     , deprecatedOpt [t1] $ tmOpt cvc4TimeoutOLD
+     , copyOpt (const $ configOptionText cvc5StrictParsing) strictSMTParseOpt
      ] <> SMT2.smtlib2Options
 
-cvc4Adapter :: SolverAdapter st
-cvc4Adapter =
+cvc5Adapter :: SolverAdapter st
+cvc5Adapter =
   SolverAdapter
-  { solver_adapter_name = "cvc4"
-  , solver_adapter_config_options = cvc4Options
-  , solver_adapter_check_sat = runCVC4InOverride
-  , solver_adapter_write_smt2 = writeCVC4SMT2File
+  { solver_adapter_name = "cvc5"
+  , solver_adapter_config_options = cvc5Options
+  , solver_adapter_check_sat = runCVC5InOverride
+  , solver_adapter_write_smt2 = writeCVC5SMT2File
   }
 
 indexType :: [SMT2.Sort] -> SMT2.Sort
 indexType [i] = i
-indexType il = SMT2.smtlib2StructSort @CVC4 il
+indexType il = SMT2.smtlib2StructSort @CVC5 il
 
 indexCtor :: [SMT2.Term] -> SMT2.Term
 indexCtor [i] = i
-indexCtor il = SMT2.smtlib2StructCtor @CVC4 il
+indexCtor il = SMT2.smtlib2StructCtor @CVC5 il
 
-instance SMT2.SMTLib2Tweaks CVC4 where
-  smtlib2tweaks = CVC4
+instance SMT2.SMTLib2Tweaks CVC5 where
+  smtlib2tweaks = CVC5
 
   smtlib2arrayType il r = SMT2.arraySort (indexType il) r
 
@@ -139,22 +126,26 @@ instance SMT2.SMTLib2Tweaks CVC4 where
   smtlib2StructCtor args = Syntax.term_app "mkTuple" args
   smtlib2StructProj _n i x = Syntax.term_app (Syntax.builder_list ["_", "tupSel", fromString (show i)]) [ x ]
 
-cvc4Features :: ProblemFeatures
-cvc4Features = useComputableReals
+cvc5Features :: ProblemFeatures
+cvc5Features = useComputableReals
            .|. useIntegerArithmetic
            .|. useSymbolicArrays
            .|. useStrings
            .|. useStructs
            .|. useFloatingPoint
+           .|. useUnsatCores
+           .|. useUnsatAssumptions
+           .|. useUninterpFunctions
+           .|. useDefinedFunctions
            .|. useBitvectors
            .|. useQuantifiers
 
-writeMultiAsmpCVC4SMT2File
+writeMultiAsmpCVC5SMT2File
    :: ExprBuilder t st fs
    -> Handle
    -> [BoolExpr t]
    -> IO ()
-writeMultiAsmpCVC4SMT2File sym h ps = do
+writeMultiAsmpCVC5SMT2File sym h ps = do
   bindings <- getSymbolVarBimap sym
   out_str  <- Streams.encodeUtf8 =<< Streams.handleToOutputStream h
   in_str <- Streams.nullInput
@@ -162,27 +153,27 @@ writeMultiAsmpCVC4SMT2File sym h ps = do
   strictness <- maybe Strict
                 (\c -> if fromConcreteBool c then Strict else Lenient) <$>
                 (getOption =<< getOptionSetting RSP.strictSMTParsing cfg)
-  c <- SMT2.newWriter CVC4 out_str in_str nullAcknowledgementAction strictness "CVC4"
-         True cvc4Features True bindings
+  c <- SMT2.newWriter CVC5 out_str in_str nullAcknowledgementAction strictness "CVC5"
+         True cvc5Features True bindings
   SMT2.setLogic c Syntax.allLogic
   SMT2.setProduceModels c True
   forM_ ps $ SMT2.assume c
   SMT2.writeCheckSat c
   SMT2.writeExit c
 
-writeCVC4SMT2File
+writeCVC5SMT2File
    :: ExprBuilder t st fs
    -> Handle
    -> [BoolExpr t]
    -> IO ()
-writeCVC4SMT2File sym h ps = writeMultiAsmpCVC4SMT2File sym h ps
+writeCVC5SMT2File sym h ps = writeMultiAsmpCVC5SMT2File sym h ps
 
-instance SMT2.SMTLib2GenericSolver CVC4 where
-  defaultSolverPath _ = findSolverPath cvc4Path . getConfiguration
+instance SMT2.SMTLib2GenericSolver CVC5 where
+  defaultSolverPath _ = findSolverPath cvc5Path . getConfiguration
 
   defaultSolverArgs _ sym = do
     let cfg = getConfiguration sym
-    timeout <- getOption =<< getOptionSetting cvc4Timeout cfg
+    timeout <- getOption =<< getOptionSetting cvc5Timeout cfg
     let extraOpts = case timeout of
                       Just (ConcreteInteger n) | n > 0 -> ["--tlimit-per=" ++ show n]
                       _ -> []
@@ -190,60 +181,60 @@ instance SMT2.SMTLib2GenericSolver CVC4 where
 
   getErrorBehavior _ = SMT2.queryErrorBehavior
 
-  defaultFeatures _ = cvc4Features
+  defaultFeatures _ = cvc5Features
 
   supportsResetAssertions _ = True
 
   setDefaultLogicAndOptions writer = do
-    -- Tell CVC4 to use all supported logics.
+    -- Tell cvc5 to use all supported logics.
     SMT2.setLogic writer Syntax.allLogic
-    -- Tell CVC4 to produce models
+    -- Tell cvc5 to produce models
     SMT2.setProduceModels writer True
 
-runCVC4InOverride
+runCVC5InOverride
   :: ExprBuilder t st fs
   -> LogData
   -> [BoolExpr t]
   -> (SatResult (GroundEvalFn t, Maybe (ExprRangeBindings t)) () -> IO a)
   -> IO a
-runCVC4InOverride = SMT2.runSolverInOverride CVC4 nullAcknowledgementAction
-                    (SMT2.defaultFeatures CVC4) (Just cvc4StrictParsing)
+runCVC5InOverride = SMT2.runSolverInOverride CVC5 nullAcknowledgementAction
+                    (SMT2.defaultFeatures CVC5) (Just cvc5StrictParsing)
 
--- | Run CVC4 in a session. CVC4 will be configured to produce models, but
+-- | Run cvc5 in a session. cvc5 will be configured to produce models, but
 -- otherwise left with the default configuration.
-withCVC4
+withCVC5
   :: ExprBuilder t st fs
   -> FilePath
-    -- ^ Path to CVC4 executable
+    -- ^ Path to cvc5 executable
   -> LogData
-  -> (SMT2.Session t CVC4 -> IO a)
+  -> (SMT2.Session t CVC5 -> IO a)
     -- ^ Action to run
   -> IO a
-withCVC4 = SMT2.withSolver CVC4 nullAcknowledgementAction
-           (SMT2.defaultFeatures CVC4) (Just cvc4StrictParsing)
+withCVC5 = SMT2.withSolver CVC5 nullAcknowledgementAction
+           (SMT2.defaultFeatures CVC5) (Just cvc5StrictParsing)
 
 setInteractiveLogicAndOptions ::
   SMT2.SMTLib2Tweaks a =>
   WriterConn t (SMT2.Writer a) ->
   IO ()
 setInteractiveLogicAndOptions writer = do
-    -- Tell CVC4 to acknowledge successful commands
+    -- Tell cvc5 to acknowledge successful commands
     SMT2.setOption writer "print-success"  "true"
-    -- Tell CVC4 to produce models
+    -- Tell cvc5 to produce models
     SMT2.setOption writer "produce-models" "true"
-    -- Tell CVC4 to make declarations global, so they are not removed by 'pop' commands
+    -- Tell cvc5 to make declarations global, so they are not removed by 'pop' commands
     SMT2.setOption writer "global-declarations" "true"
-    -- Tell CVC4 to compute UNSAT cores, if that feature is enabled
+    -- Tell cvc5 to compute UNSAT cores, if that feature is enabled
     when (supportedFeatures writer `hasProblemFeature` useUnsatCores) $ do
       SMT2.setOption writer "produce-unsat-cores" "true"
-    -- Tell CVC4 to use all supported logics.
+    -- Tell cvc5 to use all supported logics.
     SMT2.setLogic writer Syntax.allLogic
 
-instance OnlineSolver (SMT2.Writer CVC4) where
+instance OnlineSolver (SMT2.Writer CVC5) where
   startSolverProcess feat mbIOh sym = do
     timeout <- SolverGoalTimeout <$>
-               (getOpt =<< getOptionSetting cvc4Timeout (getConfiguration sym))
-    SMT2.startSolver CVC4 SMT2.smtAckResult setInteractiveLogicAndOptions
-          timeout feat (Just cvc4StrictParsing) mbIOh sym
+               (getOpt =<< getOptionSetting cvc5Timeout (getConfiguration sym))
+    SMT2.startSolver CVC5 SMT2.smtAckResult setInteractiveLogicAndOptions
+          timeout feat (Just cvc5StrictParsing) mbIOh sym
 
-  shutdownSolverProcess = SMT2.shutdownSolver CVC4
+  shutdownSolverProcess = SMT2.shutdownSolver CVC5